workers/func1-python/requirements.txt
conductor-python==1.3.10
opentelemetry-api==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
prometheus-client==0.21.1
